Pipelining finally working (almost but shh) :pray:
Selective Repeat is now working but the WINDOW_SIZE has to start at 1.
Otherwise, only the last packet is sent. (if the size = 3, only the third packet is sent).
Each ACK received increases the window size.

The app can now send the number of int that she wants without telling the protocol the number of packets before.